+++ appGUI/MainGUI.py
@@ -4052,13 +4052,12 @@ class MainGUI(QtWidgets.QMainWindow):
                         # elif self.app.exc_editor.active_tool.name == 'slot_array' \
                         #         and self.app.exc_editor.active_tool.sarray_tool.length != 0.0:
                         #     pass
-                        elif self.app.exc_editor.active_tool.name == 'move' \
+                        elif self.app.exc_editor.active_tool.name == 'drill_move' \
                                 and self.app.exc_editor.active_tool.move_tool.length != 0.0 \
                                 and self.app.exc_editor.active_tool.move_tool.width != 0.0:
                             pass
-                        elif self.app.exc_editor.active_tool.name == 'copy' \
-                                and self.app.exc_editor.active_tool.copy_tool.length != 0.0 \
-                                and self.app.exc_editor.active_tool.copy_tool.width != 0.0:
+                        elif self.app.exc_editor.active_tool.name == 'drill_copy' \
+                                and self.app.exc_editor.active_tool.copy_tool.length != 0.0:
                             pass
                         else:
                             curr_pos = self.app.exc_editor.snap_x, self.app.exc_editor.snap_y